requirements.txt
appdirs==1.4.3
argon2-cffi==19.2.0
asgiref==3.2.3
astroid==2.4.1
atomicwrites==1.3.0
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
autopep8==1.5.3
bcrypt==3.1.7
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
colorama==0.4.3
conf==0.4.1
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
daphne==2.5.0
decorator==4.4.2
Django==3.0.3
django-bootstrap3==12.0.3
django-braces==1.14.0
django-cors-headers==3.2.1
django-datetimepicker==3.14
django-filter==2.2.0
django-mysql==3.3.0
django-softdelete==0.9.1
django-tempus-dominus==5.1.2.13
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
fake-useragent==0.1.11
Faker==4.0.2
hyperlink==19.0.0
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
incremental==17.5.0
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.2.1
matplotlib==3.2.1
mccabe==0.6.1
misaka==2.1.1
more-itertools==8.2.0
moviepy==1.0.3
mysqlclient==1.4.6
networkx==2.4
numpy==1.18.2
opencv-python==4.2.0.34
packaging==20.3
pandas==1.0.3
paramiko==2.7.1
parse==1.15.0
pep8==1.7.1
Pillow==7.0.0
pipreqs==0.4.10
pluggy==0.13.1
proglog==0.1.9
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoIt==0.4
pycodestyle==2.6.0
pycparser==2.20
pyee==7.0.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.2
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyppeteer==0.0.25
pyquery==1.4.1
pysftp==0.2.9
PySide2==5.14.2.1
pytest==5.4.1
python-dateutil==2.8.1
python-qt5==0.1.10
pytz==2019.3
qtgui==0.0.1
qThread==0.1.0
range-regex==1.0.4
requests==2.23.0
requests-html==0.10.0
rpn-calc==0.1.5
schedule==0.5.0
scp==0.13.2
selenium==3.13.0
service-identity==18.1.0
shiboken2==5.14.2.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.1
tqdm==4.45.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
w3lib==1.21.0
wcwidth==0.1.9
websockets==8.1
wrapt==1.12.1
yarg==0.1.9
zope.interface==5.1.0
